refactor: move to internal utils
