+++ src/utils.ts
@@ -67,24 +67,3 @@ export function getProtocol(id: string): string | undefined {
   const proto = id.match(ProtocolRegex);
   return proto ? proto.groups.proto : undefined;
 }
-
-export function clearImports(imports: string) {
-  return (imports || "")
-    .replace(/(\/\/[^\n]*\n|\/\*.*\*\/)/g, "")
-    .replace(/\s+/g, " ");
-}
-
-export function getImportNames(cleanedImports: string) {
-  const topLevelImports = cleanedImports.replace(/{([^}]*)}/, "");
-  const namespacedImport = topLevelImports.match(/\* as \s*(\S*)/)?.[1];
-  const defaultImport =
-    topLevelImports
-      .split(",")
-      .find((index) => !/[*{}]/.test(index))
-      ?.trim() || undefined;
-
-  return {
-    namespacedImport,
-    defaultImport,
-  };
-}
